First version of combined budget
