+++ src/views/categories/AddCategory.tsx
@@ -6,14 +6,14 @@ import { CategoryForm } from '../../components/categories/CategoryForm';
 import { btApp } from '../../BudgetTracker';
 import { goBack } from '../../domain/utils/goBack';
 import { CategoryPaths } from '../../domain/paths/CategoryPaths';
-import { CloseButton } from '../../components/buttons/CloseButton';
+import { CloseButtonHistory } from '../../components/buttons/CloseButton';
 
 export const AddCategory: React.FC<RouterProps&HeaderNotifierProps> = (props) => {
     
     React.useEffect(() => {
         
         props.onTitleChange('Add category');
-        props.onActions(<CloseButton history={props.history}/>);
+        props.onActions(<CloseButtonHistory history={props.history}/>);
         return function () {
             props.onActions(undefined);
             props.onTitleChange('');
