+++ src/views/expenses/ExpenseAdd.tsx
@@ -9,7 +9,7 @@ import { DateDay } from "../../domain/DateDay";
 import { uuid } from "../../domain/utils/uuid";
 import { useBudgetModel } from "../../hooks/useBudgetModel";
 import { useCurrentCountry } from "../../hooks/useCurrentCountry";
-import { CloseButton } from "../../components/buttons/CloseButton";
+import { CloseButtonHistory } from "../../components/buttons/CloseButton";
 import { ExpenseForm } from "../../components/expenses/ExpenseForm";
 import CircularProgress from "@material-ui/core/CircularProgress";
 
@@ -49,7 +49,7 @@ export const ExpenseAdd: React.FC<ExpenseViewProps> = (props) => {
 
     React.useEffect(()=> {
         onTitleChange('Add expense');
-        onActions([<CloseButton history={history} key='close-button'/>]);
+        onActions([<CloseButtonHistory history={history} key='close-button'/>]);
         return function () {
             onActions(null); 
         }
