+++ tsconfig.json
@@ -7,6 +7,7 @@
       "esnext"
     ],
     "allowJs": true,
+    "downlevelIteration": true,
     "skipLibCheck": true,
     "esModuleInterop": true,
     "allowSyntheticDefaultImports": true,
